html/semantics/scripting-1/the-script-element/scriptexecute-events/009.html

<!DOCTYPE html>
<html>
  <title>Removing script in beforescriptexecute</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    var targets = [];
    var t = async_test(function() {
      document.addEventListener("beforescriptexecute", this.step_func(function(e) {
        targets.push(e.target.id);
        e.target.parentNode.removeChild(e.target);
      }), false);
    });
  </script>
  <script id=last>
    t.step(function() {
      assert_array_equals(targets, ["last"]);
    });
    t.done();
  </script>
